refactored provideInstruction method in Succulents.java and refactored Condition.java

+++ src/main/java/com/plantTracker/plantTracker/models/enums/Condition.java
@@ -1,12 +1,12 @@
 package com.plantTracker.plantTracker.models.enums;
 
 public class Condition {
-    private double minTemperature;
-    private double maxTemperature;
-    private double minHumidity;
-    private double maxHumidity;
+    private int minTemperature;
+    private int maxTemperature;
+    private int minHumidity;
+    private int maxHumidity;
 
-    public Condition(double minTemperature, double maxTemperature, double minHumidity, double maxHumidity) {
+    public Condition(int minTemperature, int maxTemperature, int minHumidity, int maxHumidity) {
         this.minTemperature = minTemperature;
         this.maxTemperature = maxTemperature;
         this.minHumidity = minHumidity;
@@ -16,35 +16,35 @@ public class Condition {
     public Condition() {
     }
 
-    public double getMinTemperature() {
+    public int getMinTemperature() {
         return minTemperature;
     }
 
-    public void setMinTemperature(double minTemperature) {
+    public void setMinTemperature(int minTemperature) {
         this.minTemperature = minTemperature;
     }
 
-    public double getMaxTemperature() {
+    public int getMaxTemperature() {
         return maxTemperature;
     }
 
-    public void setMaxTemperature(double maxTemperature) {
+    public void setMaxTemperature(int maxTemperature) {
         this.maxTemperature = maxTemperature;
     }
 
-    public double getMinHumidity() {
+    public int getMinHumidity() {
         return minHumidity;
     }
 
-    public void setMinHumidity(double minHumidity) {
+    public void setMinHumidity(int minHumidity) {
         this.minHumidity = minHumidity;
     }
 
-    public double getMaxHumidity() {
+    public int getMaxHumidity() {
         return maxHumidity;
     }
 
-    public void setMaxHumidity(double maxHumidity) {
+    public void setMaxHumidity(int maxHumidity) {
         this.maxHumidity = maxHumidity;
     }
 }
